Server: Add support for sidebar in user pages
